refactor: move func to generate random string to k8s storage
